refactor: improve navigation using NavigationLink value
Streamlines navigation by using NavigationLink's value parameter and navigationDestination modifiers.

This change removes the need to explicitly pass astronauts to MissionsGrid and MissionsList, simplifying the view structure. It also introduces Hashable conformance to Astronaut, Mission and CrewMember to allow them to be passed as NavigationLink values.

Also improve the performance since Views are only loaded when are called